pyproject.toml
[project]
name = "smart-organaizer"
version = "0.1.0"
description = ""
authors = [
    {name = "chichxh",email = "egorodand@gmail.com"}
]
readme = "README.md"
package-mode = false
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain (>=0.3,<0.4)",
    "langgraph (>=0.2.36,<0.3.0)",
    "langchain-gigachat (>=0.3.12,<0.4.0)",
    "google-api-python-client (>=2.140.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytz (>=2025.2,<2026.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
